sql: coerce all numbers to bigints

+++ src/sql/addressUnspents.sql
@@ -1,8 +1,8 @@
 SELECT
 tx_hash AS "txId",
-confirmations AS "confirmations",
+confirmations::bigint AS "confirmations",
 addr_bs58 AS "address",
-txout_value AS "value",
-txout_pos AS "vout"
+txout_value::bigint AS "value",
+txout_pos::bigint AS "vout"
 FROM unspent_txout_view
 WHERE addr_bs58 IN ({{addresses | join(',')}})
